refactor: product Option 관련 폴더구조 변경
